requirements.txt
# Core Flask
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
Flask-CORS==5.0.0

# Trading: market data
yfinance>=0.2.40

# Trading: AI trader (Gemini API)
google-genai>=1.0.0

# Trading: scheduled runs
APScheduler>=3.10.0
